Move useAnnotationsModal to console-shared
