+++ frontend/public/components/utils/details-page.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next';
 import * as _ from 'lodash-es';
 import { Button } from '@patternfly/react-core';
 import { PencilAltIcon } from '@patternfly/react-icons';
-import { useCanClusterUpgrade } from '@console/shared';
-import { useAnnotationsModal } from '@console/dynamic-plugin-sdk/src/app/components/modals/useAnnotationsModal';
+import { useCanClusterUpgrade } from '@console/shared/src/hooks/useCanClusterUpgrade';
+import { useAnnotationsModal } from '@console/shared/src/hooks/useAnnotationsModal';
 import { DetailsItem } from './details-item';
 import { Kebab } from './kebab';
 import { LabelList } from './label-list';
