+++ frontend/public/reducers/features.ts
@@ -24,8 +24,7 @@ import {
   MachineModel,
   PrometheusModel,
 } from '../models';
-import { referenceForGroupVersionKind } from '../module/k8s';
-import { referenceForModel } from '../module/k8s/k8s-ref';
+import { referenceForGroupVersionKind, referenceForModel } from '../module/k8s/k8s-ref';
 import { RootState } from '../redux';
 import { ActionType as K8sActionType } from '@console/dynamic-plugin-sdk/src/app/k8s/actions/k8s';
 import { FeatureState } from '@console/dynamic-plugin-sdk/src/app/features';
